remove dependence on starlette and fastapi
