+++ src/_canary/resource_pool/__init__.py
@@ -1,4 +1,3 @@
 from .adapter import ResourcePoolAdapter  # noqa: F401
 from .rpool import ResourcePool  # noqa: F401
-from .startup import make_resource_pool  # noqa: F401
-from .startup import start_resource_pool_server  # noqa: F401
+from .rpool import make_resource_pool  # noqa: F401
